requirements.txt
python-dotenv
requests
streamlit
streamlit_drawable_canvas
watchdog